initial commit to the branch with much of the editor functionality removed for now while I remove bugs
