+++ editor.py
@@ -1,0 +1,51 @@
+import curses
+
+
+from npc import NPC
+
+
+def draw_menu(stdscr, selected_row_idx):
+    stdscr.clear()
+    h, w = stdscr.getmaxyx()
+
+    title = "Editor Menu"
+    stdscr.addstr(0, w//2 - len(title)//2, title)
+
+    menu_items = ["Create Location", "Create Enemy", "Exit"]
+    for idx, item in enumerate(menu_items):
+        x = w//2 - len(item)//2
+        y = h//2 - len(menu_items)//2 + idx
+        if idx == selected_row_idx:
+            stdscr.attron(curses.A_REVERSE)
+            stdscr.addstr(y, x, item)
+            stdscr.attroff(curses.A_REVERSE)
+        else:
+            stdscr.addstr(y, x, item)
+
+    stdscr.refresh()
+
+def main(stdscr):
+    curses.curs_set(0)
+    curses.init_pair(1, curses.COLOR_BLACK, curses.COLOR_WHITE)
+    stdscr.clear()
+    stdscr.refresh()
+
+    current_row = 0
+
+    while True:
+        draw_menu(stdscr, current_row)
+        key = stdscr.getch()
+
+        if key == curses.KEY_UP and current_row > 0:
+            current_row -= 1
+        elif key == curses.KEY_DOWN and current_row < 2:
+            current_row += 1
+        elif key == curses.KEY_ENTER or key in [10, 13]:
+            if current_row == 0:
+                create_location(stdscr)
+            elif current_row == 1:
+                create_enemy(stdscr)
+            elif current_row == 2:
+                break
+
+curses.wrapper(main)
